requirements.txt
fastapi
uvicorn
pydantic
python-jose
passlib
